separated bondInfo and fixed offerings
